bugfix + use of textures instead of instanceArray because it was not working on mobile otherwise + meshcap texture atlas to power of 2 size + no post processing in mobile on chibi example + forgot to read transform matrix on .mcap parse

+++ src/tracking/util/face-tracker-utils.ts
@@ -36,5 +36,5 @@ export const createFaceLandmarksIndexAttribute = ( mesh:Mesh ) =>
 			posIndex2LandmarkIndex.push(idx < FACE_LANDMARKS_COUNT ? idx : 65535);
         } 
 
-		geometry.setAttribute("landmarkIndex", new BufferAttribute(new Uint16Array(posIndex2LandmarkIndex), 1));	
+		geometry.setAttribute("landmarkIndex", new BufferAttribute(new Float32Array(posIndex2LandmarkIndex), 1));	
 }
